Add endpoints and basic main

+++ backend/dabox-core/src/entity.rs
@@ -1,6 +1,6 @@
 /// This module provides an `Entity` trait that is used to identify entities in the system.
 /// An entity has a unique identifier that is used to identify it in the system (i.e managing directory permissions).
-pub type EntityUid = i32;
+pub type EntityUid = u64;
 
 pub trait Entity {
     fn uid(&self) -> EntityUid;
